tests/wpt/tests/css/css-ui/parsing/outline-offset-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().outlineOffset</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline-offset">
<meta name="assert" content="outline-offset computed value is absolute length.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<style>
  #target {
    font-size: 40px;
  }
</style>
<div id="target"></div>
<script>
test_computed_value("outline-offset", "2.5px", ["2.5px", "2px"]); // See https://github.com/w3c/csswg-drafts/issues/12906
test_computed_value("outline-offset", "10px");
test_computed_value("outline-offset", "0.5em", "20px");
test_computed_value("outline-offset", "calc(10px + 0.5em)", "30px");
test_computed_value("outline-offset", "calc(10px - 0.5em)", "-10px");

test(() => {
  target.style['outline-offset'] = '10px';
  target.style['outline-style'] = 'none';
  assert_equals(getComputedStyle(target)['outline-offset'], '10px');
  target.style['outline-style'] = 'auto';
  assert_equals(getComputedStyle(target)['outline-offset'], '10px');
  target.style['outline-style'] = '';
}, 'outline-offset is independent of the value of outline-style');
</script>
</body>
</html>
